+++ simple_history/tests/admin.py
@@ -1,3 +1,5 @@
+from __future__ import unicode_literals
+
 from django.contrib import admin
 
 from simple_history.admin import SimpleHistoryAdmin