+++ simple_history/tests/tests.py
@@ -1,10 +1,16 @@
+from __future__ import unicode_literals
+
 from datetime import datetime, timedelta
 from django import VERSION
 from django.test import TestCase
 from django_webtest import WebTest
 from django.core.files.base import ContentFile
 from django.core.urlresolvers import reverse
-from django.contrib.auth.models import User
+try:
+    from django.contrib.auth import get_user_model
+    User = get_user_model()
+except ImportError:  # django 1.4 compatibility
+    from django.contrib.auth.models import User
 
 from .models import Poll, Choice, Restaurant, Person, FileModel, Document, Book, Library, State
 from .models import ExternalModel1, ExternalModel3
@@ -335,7 +341,7 @@ def get_history_url(model, history_index=None):
 
 class AdminSiteTest(WebTest):
     def setUp(self):
-        self.user = User.objects.create_superuser('u', 'u@example.com', 'pass')
+        self.user = User.objects.create_superuser('user_login', 'u@example.com', 'pass')
 
     def login(self, user=None):
         if user is None:
@@ -346,12 +352,17 @@ class AdminSiteTest(WebTest):
         return form.submit()
 
     def test_history_list(self):
+        if VERSION >= (1, 5):
+            self.assertEqual(self.user._meta.module_name, 'customuser')
         self.login()
-        poll = Poll.objects.create(question="why?", pub_date=today)
+        poll = Poll(question="why?", pub_date=today)
+        poll._history_user = self.user
+        poll.save()
         response = self.app.get(get_history_url(poll))
-        self.assertIn(get_history_url(poll, 0), response.content)
-        self.assertIn("Poll object", response.content)
-        self.assertIn("Created", response.content)
+        self.assertIn(get_history_url(poll, 0), response.unicode_normal_body)
+        self.assertIn("Poll object", response.unicode_normal_body)
+        self.assertIn("Created", response.unicode_normal_body)
+        self.assertIn(self.user.username, response.unicode_normal_body)
 
     def test_history_form_permission(self):
         self.login(self.user)
@@ -365,7 +376,7 @@ class AdminSiteTest(WebTest):
         response.form['question'] = ""
         response = response.form.submit()
         self.assertEqual(response.status_code, 200)
-        self.assertIn("This field is required", response.content)
+        self.assertIn("This field is required", response.unicode_normal_body)
 
     def test_history_form(self):
         self.login()